+++ scanner/scan.go
@@ -1,4 +1,4 @@
-package make
+package scanner
 
 import (
 	"bytes"
