+++ scanner/scanner_suite_test.go
@@ -1,0 +1,13 @@
+package scanner_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestScanner(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Scanner Suite")
+}
